add Sigmoid layer, arrange module structure
