+++ src/dlfs01/common/layers/mod.rs
@@ -1,0 +1,6 @@
+//! layers
+//!
+//! Layers
+
+pub mod activation;
+pub mod layer_base;
