+++ src/main.rs
@@ -31,7 +31,7 @@ fn main() {
 
     // commom
     // rs_deep::dlfs01::common::loss_function::main();
-    rs_deep::dlfs01::common::layers::main();
+    rs_deep::dlfs01::common::layers::activation::main();
 
     // dataset
     // rs_deep::dlfs01::dataset::main();
